% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TLU.R
\name{calc_tlu}
\alias{calc_tlu}
\title{based on FAO http://www.fao.org/wairdocs/ilri/x5443e/x5443e04.htm}
\usage{
calc_tlu(df, cow_var = cows, camel_var = camels, sheep_var = sheep,
  goat_var = goats, equine_var = equines, horse_var = NA, mule_var = NA,
  ass_var = NA, pig_var = NA, chicken_var = chickens, duck_var = ducks)
}
\description{
based on FAO http://www.fao.org/wairdocs/ilri/x5443e/x5443e04.htm
}
